Move and rename deconz base entity to separate module
