+++ homeassistant/components/deconz/deconz_event.py
@@ -25,7 +25,7 @@ from homeassistant.helpers import device_registry as dr
 from homeassistant.util import slugify
 
 from .const import ATTR_DURATION, ATTR_ROTATION, CONF_ANGLE, CONF_GESTURE, LOGGER
-from .deconz_device import DeconzBase
+from .entity import DeconzBase
 from .hub import DeconzHub
 
 CONF_DECONZ_EVENT = "deconz_event"
