+++ homeassistant/components/deconz/sensor.py
@@ -56,7 +56,7 @@ from homeassistant.helpers.typing import StateType
 import homeassistant.util.dt as dt_util
 
 from .const import ATTR_DARK, ATTR_ON
-from .deconz_device import DeconzDevice
+from .entity import DeconzEntity
 from .hub import DeconzHub
 
 PROVIDES_EXTRA_ATTRIBUTES = (
@@ -390,7 +390,7 @@ async def async_setup_entry(
     )
 
 
-class DeconzSensor(DeconzDevice[SensorResources], SensorEntity):
+class DeconzSensor(DeconzEntity[SensorResources], SensorEntity):
     """Representation of a deCONZ sensor."""
 
     TYPE = SENSOR_DOMAIN
